after consulting and starting filters
